# src/pkg/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for authenticating and authorizing Kubernetes access via Tailscale identity.",
        "title": "Tailscale Kubernetes Auth API",
        "contact": {
            "name": "Specht Labs",
            "url": "specht-labs.de",
            "email": "tka@specht-labs.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "tka.sphinx-map.ts.net:8123",
    "basePath": "/api/v1alpha1",
    "paths": {
        "/api/v1alpha1/kubeconfig": {
            "get": {
                "security": [
                    {
                        "TailscaleAuth": []
                    }
                ],
                "description": "Generates and returns a kubeconfig file for the authenticated Tailscale user",
                "produces": [
                    "application/yaml",
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Get kubeconfig for authenticated user",
                "responses": {
                    "200": {
                        "description": "OK - Returns kubeconfig file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Tagged nodes not supported or error unmarshaling capability or multiple capability rules",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Request from Funnel or no capability rule found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - User not authenticated or credentials not ready",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Error with WhoIs or generating kubeconfig",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1alpha1/login": {
            "get": {
                "security": [
                    {
                        "TailscaleAuth": []
                    }
                ],
                "description": "Retrieves the current authentication status for a Tailscale user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Get user authentication status",
                "responses": {
                    "200": {
                        "description": "OK - Returns the current user authentication status",
                        "schema": {
                            "$ref": "#/definitions/models.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Tagged nodes not supported or error unmarshaling capability or multiple capability rules",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Request from Funnel or no capability rule found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Error with WhoIs or retrieving user status",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "tailscaleAuth": []
                    }
                ],
                "description": "Authenticates a user through Tailscale, validates their capability rule, and provisions Kubernetes credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Authenticate user and provision Kubernetes credentials",
                "responses": {
                    "202": {
                        "description": "Accepted - User authenticated and credentials are being provisioned",
                        "schema": {
                            "$ref": "#/definitions/models.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Tagged nodes not supported or error unmarshaling capability or multiple capability rules",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Request from Funnel or no capability rule found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - Invalid capability rule (period too short)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Error with WhoIs, parsing duration, or signing in user",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1alpha1/logout": {
            "post": {
                "security": [
                    {
                        "TailscaleAuth": []
                    }
                ],
                "description": "Revokes Kubernetes credentials for the authenticated Tailscale user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Log out authenticated user",
                "responses": {
                    "200": {
                        "description": "OK - User successfully logged out with login info",
                        "schema": {
                            "$ref": "#/definitions/models.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Tagged nodes not supported or error unmarshaling capability or multiple capability rules",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Request from Funnel or no capability rule found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Error with WhoIs, parsing duration, or during logout process",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "description": "Structured error response with contextual advice",
            "type": "object",
            "properties": {
                "advice": {
                    "description": "List of suggestions to help resolve the error\nexample: [\"Check your Tailscale connection\", \"Verify you have the required capabilities\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Primary error message\nexample: Failed to authenticate user",
                    "type": "string"
                }
            }
        },
        "models.UserLoginResponse": {
            "description": "Contains authenticated user information and session details",
            "type": "object",
            "properties": {
                "role": {
                    "description": "Role assigned to the user in Kubernetes\nexample: cluster-admin",
                    "type": "string"
                },
                "until": {
                    "description": "Expiration timestamp of the authentication credentials in RFC3339 format\nexample: 2023-12-31T23:59:59Z",
                    "type": "string"
                },
                "username": {
                    "description": "Username of the authenticated user\nexample: alice@example.com",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "TailscaleAuth": {
            "description": "Authentication happens automatically via the Tailscale network. The server performs a WhoIs lookup on the client's IP address to determine identity. This header is for documentation purposes only and is not actually required to be set.",
            "type": "apiKey",
            "name": "X-Tailscale-User",
            "in": "header"
        }
    }
}